+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Ember Flexberry Changelog
 ### New Features & improvements
 
+
+## 2016-06-24
+### Breaking changes
+ObjectlistviewComponent:
+* Template is split for desktop and mobile phones.
+* Remove properties `useSingleColumn` & `emptyMobileHeader`.
+
 ## 2016-06-23
 ### Added
 FlexberryObjectlistviewComponent